feat: Add ODS Calls and CallChargesPerYear package and db script changes.

+++ db_scripts/create_table_ods_callchargesperyear.sql
@@ -1,7 +1,7 @@
 USE [SSIS_PROJECT_ODS]
 GO
 
-/****** Object:  Table [dbo].[CallChargesPerYear]    Script Date: 09/06/2022 16:33:56 ******/
+/****** Object:  Table [dbo].[CallChargesPerYear]    Script Date: 10/06/2022 14:10:56 ******/
 SET ANSI_NULLS ON
 GO
 
@@ -9,8 +9,8 @@ SET QUOTED_IDENTIFIER ON
 GO
 
 CREATE TABLE [dbo].[CallChargesPerYear](
-	[CallTypeKey] [int] NULL,
-	[Year] [bigint] NULL,
+	[CallTypeKey] [smallint] NULL,
+	[Year] [int] NULL,
 	[CallChargesPerMin] [real] NULL,
 	[CallType] [nvarchar](50) NULL
 ) ON [PRIMARY]
